% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{all_equal_}
\alias{all_equal_}
\title{Returns all.equal strictly with a logical}
\usage{
all_equal_(...)
}
\arguments{
\item{...}{Any Parameters passed to base all.equal}
}
\value{
Logical
}
\description{
Returns True if target and current match, False otherwise. Including
parameters for all.equal {base}
}
\examples{
all_equal_(3, 3)
all_equal_(1:5, 1:5)
all_equal_(1.1:5.1, 1:5)
all_equal_(1.1:5.1, 1:5, tolerance = 0.2)
}
\seealso{
\code{\link{all.equal}} is the same function, but the base function
  can return more than logical.
}

