% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_functions.R
\name{combo_chr}
\alias{combo_chr}
\title{Combinations of character vector}
\usage{
combo_chr(.set)
}
\arguments{
\item{.set}{Character vector to combine}
}
\value{
List of vector of character combinations
}
\description{
Returns a map of all non-empty combinations of set. Returns an empty list if
given an empty or invalid value.
}
\examples{
combo_chr(letters[1:5])
}

