% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_functions.R
\name{combo_dbl}
\alias{combo_dbl}
\title{Combinations of double vector}
\usage{
combo_dbl(.set)
}
\arguments{
\item{.set}{Double vector to combine}
}
\value{
List of vector of double combinations
}
\description{
Returns a map of all non-empty combinations of set. Returns an empty list if
given an empty or invalid value.
}
\examples{
combo_dbl(1.5:5.5)
}

