% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_functions.R
\name{combo_internal}
\alias{combo_internal}
\title{Provide combinations of vector elements}
\usage{
combo_internal(.set, .map_fn)
}
\arguments{
\item{.set}{Vector of values to combine}

\item{.map_fn}{Function to map over .set elements}
}
\value{
List of vectors (chr, lgl, int, dbl) that match set type
}
\description{
Returns a map of all non-empty combinations of set. Returns an empty list if
given an empty or invalid value.
}

