% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_functions.R
\name{combo_list}
\alias{combo_list}
\title{Combinations of list elements}
\usage{
combo_list(.set)
}
\arguments{
\item{.set}{List of values to combine}
}
\value{
List of list of values
}
\description{
Returns a map of all non-empty combinations of set. Returns an empty list if
given an empty or invalid value.
}
\examples{
combo_list(list(list(1:5),list(6:10),list(11:15)))
}

