% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_to_r_code_functions.R
\name{dataframe_to_r_code}
\alias{dataframe_to_r_code}
\title{The text of R Code version of Data Frame}
\usage{
dataframe_to_r_code(df)
}
\arguments{
\item{df}{A Data Frame}
}
\value{
Character of R Code to make dataframe
}
\description{
The text of R Code version of Data Frame
}
\examples{
dataframe_to_r_code(data.frame(a=1:5))
dataframe_to_r_code(data.frame(letters=letters))
}

