% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even_or_odd_functions.R
\name{even_count_of_elements}
\alias{even_count_of_elements}
\title{Checks if the number of elements in a vector is even or not}
\usage{
even_count_of_elements(v)
}
\arguments{
\item{v}{Vector of any type}
}
\value{
logical TRUE if the vector has an Even length, FALSE otherwise
}
\description{
Checks if the number of elements in a vector is even or not
}

