% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even_or_odd_functions.R
\name{even_or_odd}
\alias{even_or_odd}
\title{Marks a numeric vector as either Even or Odd}
\usage{
even_or_odd(v)
}
\arguments{
\item{v}{Vector of numeric type}
}
\value{
character Either "Even" or "Odd"
}
\description{
If the vector has an even number of elements, then it is always Even
If the vector has an odd number of elements, but any of them are even
 then the result is Even
Otherwise, the result is "Odd"
}
\examples{
even_or_odd(1:4)
even_or_odd(1:3)
even_or_odd(1:5)
even_or_odd(2:4)
even_or_odd(2)
}

