% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_}
\alias{get_}
\title{Get an object by name}
\usage{
get_(.o, name)
}
\arguments{
\item{.o}{An object}

\item{name}{Character vector (or castable)}
}
\value{
Object
}
\description{
Flips the order of the {base} get()
}
\examples{
get_(list(b = 4:2), 'b')
}

