% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{get_col}
\alias{get_col}
\title{Select columns by indices}
\usage{
get_col(.m, cols)
}
\arguments{
\item{.m}{A 2d matrix}

\item{cols}{Numeric or character vector or empty (missing)}
}
\value{
Subset of columns of matrix.
}
\description{
Select columns by indices
}
\examples{
get_col(matrix(1:6, nrow = 2), 2)
get_col(matrix(1:6, nrow = 2), c(2,3))
get_col(matrix(1:6, nrow = 2, byrow = TRUE, dimnames = list(c("a", "b"), c("x", "y", "z"))), "z")
}

