% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{get_row}
\alias{get_row}
\title{Select rows by indices}
\usage{
get_row(.m, rows)
}
\arguments{
\item{.m}{A 2d matrix}

\item{rows}{Numeric or character vector or empty (missing)}
}
\value{
Subset of rows of matrix.
}
\description{
Select rows by indices
}
\examples{
get_row(matrix(1:6, nrow = 3), 2)
get_row(matrix(1:6, nrow = 3), c(2,3))
get_row(matrix(1:6, nrow = 3, byrow = TRUE, dimnames = list(c("a", "b", "c"), c("x", "y"))), "b")
}

