% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{if_else_}
\alias{if_else_}
\title{Apply logic based on logical}
\usage{
if_else_(.data, .truth, .lhs, .rhs)
}
\arguments{
\item{.data}{Any value}

\item{.truth}{Logical Value}

\item{.lhs}{A function}

\item{.rhs}{A function}
}
\value{
Any
}
\description{
If truth is truthy, applies lhs to data. Else, apply rhs to data
}
\examples{
if_else_(1:5, TRUE, sum, mean)
if_else_(1:5, FALSE, sum, mean)
}

