% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{is_invalid}
\alias{is_invalid}
\title{Returns if the value is invalid}
\usage{
is_invalid(.data)
}
\arguments{
\item{.data}{Any}
}
\value{
Logical
}
\description{
Returns True if NA, NULL, NaN, or of length 0
}
\examples{
is_invalid(3)
is_invalid(NULL)
is_invalid(NaN)
is_invalid(list())
}

