% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals_functions.R
\name{residuals_PRESS}
\alias{residuals_PRESS}
\title{Compute the PRESS Residuals for a class of "lm"
Large residuals in PRESS not seen in normal residuals can indicate model
overfit.}
\usage{
residuals_PRESS(fit)
}
\arguments{
\item{fit}{An object of class "lm"}
}
\value{
The PRESS residuals, \code{ei} / \code{1 - hii}.
}
\description{
Compute the PRESS Residuals for a class of "lm"
Large residuals in PRESS not seen in normal residuals can indicate model
overfit.
}
\examples{
lm(y~.,data.frame(y=1:5,x=10:14)) -> fit
residuals_PRESS(fit)
}

