% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{se}
\alias{se}
\title{Standard Errors}
\usage{
se(fit, df = NULL)
}
\arguments{
\item{fit}{Object of class "lm". A fitted model.}

\item{df}{DataFrame of predictors and predicted (optional)}
}
\value{
double Squared differences of predicted and actual values
}
\description{
Returns the standard errors of the fit againt the data.
(y - yhat)^2
Without a supplied dataframe, returns the errors from the training fit
}
\examples{
data.frame(y=1:5,x=c(1,2,2,4,5)) -> df1
data.frame(y=3:7,x=c(3,4,6,5,2)) -> df2
lm(y ~ ., df1) -> fit

se(fit,df1)
se(fit,df2)
se(fit)

}

