% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{set_colnames}
\alias{set_colnames}
\title{Set colnames of object}
\usage{
set_colnames(.m, column_names = NULL)
}
\arguments{
\item{.m}{A 2d matrix}

\item{column_names}{Numeric or character vector or empty (missing)}
}
\value{
Copy of matrix with colnames
}
\description{
Set colnames of object
}
\examples{
set_colnames(matrix(1:6, nrow = 3), c("a", "b"))
set_colnames(matrix(1:6, nrow = 2), c("a", "b", "c"))
set_colnames(matrix(1:6, nrow = 3, byrow = TRUE,
             dimnames = list(c("a", "b", "c"), c("x", "y"))), NULL)
}

