% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{set_rownames}
\alias{set_rownames}
\title{Set rownames of object}
\usage{
set_rownames(.m, row_names = NULL)
}
\arguments{
\item{.m}{A 2d matrix}

\item{row_names}{Numeric or character vector or NULL}
}
\value{
Copy of matrix with rownames
}
\description{
Set rownames of object
}
\examples{
set_rownames(matrix(1:6, nrow = 2), c("a", "b"))
set_rownames(matrix(1:6, nrow = 3), c("a", "b", "c"))
set_rownames(matrix(1:6, nrow = 3, byrow = TRUE, dimnames = list(c("a", "b", "c"))), NULL)
}

