% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_to_r_code_functions.R
\name{vector_to_r_code}
\alias{vector_to_r_code}
\title{The text of R Code version of vector}
\usage{
vector_to_r_code(v)
}
\arguments{
\item{v}{A vector}
}
\value{
Character of R Code to make vector
}
\description{
Large residuals in PRESS not seen in normal residuals can indicate model
overfit.
}
\examples{
vector_to_r_code(1:5)
vector_to_r_code(letters)
vector_to_r_code(c(T,F,F,T))
}

