% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_lm.R
\name{vif_lm}
\alias{vif_lm}
\title{Check for multicollinearity in linear model}
\usage{
vif_lm(frame, pred, bound)
}
\arguments{
\item{frame}{An object of class 'data.frame'}

\item{pred}{An string object of the response variable}

\item{bound}{An optional numeric object (default = 5)}
}
\value{
A linear model containing no multicollinearity
}
\description{
Produce a linear model in which no variables in the final model
in which no multicollinearity is present as determined by the
variance inflation factor (VIF)
}
\examples{
d <- data.frame(y=1:5,x1=sample(1:10,5),x2=sample(1:10,5),x3=sample(1:10,5))
vif_lm(d,'y',5)

vif_lm(d,'y')
}

